Ensure digitizer loss does not cause data corruption
Previously if some traces messages from a digitizer were lost, but the
digitizer continues to send data later then the frame offsets in the
`detector_data` dataset will be incorrect.

This fix ensures that missing data causes holes in the `detector_data`
and that subsequent frames are in the correct position.

This does introduce the caveat that the frame number must not reset or
overflow unless `stream-to-file` is restarted, but that is acceptable
for now.

Fixes #30.