Update ADR and docs for Rust error handling
Further restrict the use of `anyhow`.
The flexibility of this has led to some quite sloppy error handling,
particularly cases where errors that can be recovered from (e.g. only
affecting a single frame of data) are propagated to the point of being
terminal (at which point they may as well have just panicked and
provided a little more information about the error).

Explicitly list the things that should not be done and are best not to
do in the developer docs.

Note the desire to address other panicking operations and functions
later on.