Nexus Writer: Make Frame Times Relative to Run Start instead of First Received Frame (#271)
## Summary of changes

Applies to Nexus Writer.

In the cases where frames arrive out of order, it is possible for
`event_time_zero` to be 'negative'. As `event_time_zero` is unsigned
this results in integer wrap around, giving misleading results. This
follows from the `offset` attribute (which is used to calculate
`event_time_zero`) being set to the time of the first frame message.

This PR instead sets `offset` to the time of the run start. As any
frames before this time are not considered part of the run, this change
should make `event_time_zero` robust to wrap around errors.

## Instruction for review/testing

General code review.

The change has been tested and verified on simulated data.

---------

Co-authored-by: Dan Nixon <dan@dan-nixon.com>