Fixes flushed runs index bug (#280)
## Summary of changes

Bug fix.

In very particular circumstances a vector is accessed out of bounds.
When two runs are present in the run cache, and complete at the same
time the `flush` method will likely try to access the `run_cache` deque
out of bounds.

This bug was missed due to non-idiomatic code being used. It has been
replaced and a suitable test added to check this (albeit unlikely) use
case.

## Instruction for review/testing

General code review.

Tested on simulated data.