Resume Run Functionality (#273)
## Summary of changes

Added behaviour for the Nexus Writer to resume a run if it was
interrupted by the component going down.
- ~~A file with extension ".partial_run" is added to the local path when
a run_start is received. This file is removed when the nexus file is
transferred to the archive (signalling a successful run).~~
- When the nexus writer starts, any ~~".partial_run"~~ ".nxs" files in
the local path are read and runs re-created (as ~~.partial_run~~ .nxs
files indicate that a previous instance of nexus writer had failed).

## Instruction for review/testing

General code review.

Tool was tested with simulated data.