Single producer thread event formation (#285)
## Summary of changes

The current event-formation component creates a new thread to produce
each outgoing message resulting in the overhead associated with thread
for each message. In this PR, the component creates a single persistent
thread for producing messages to Kafka, and transfers processed traces
from the main thread to this thread via a channel.

## Instruction for review/testing

Commentary on the new architecture is welcome. Is this a good change?

General code review.

Has been tested on simulated and Kafka data.