Graceful shutdown (#287)
## Summary of changes

Detects a SIGINT signal from the OS and performs a graceful shutdown.

### `trace-to-events` and `digitiser-aggregator`

For trace-to-events and digitiser-aggregator, both the producer thread,
and the main thread listens for SIGINT.

- When the producer thread hears it, it closes the MPSC channel and
flushes all messages (eventlists and frames respectively) before
terminating the thread.
- When the main-thread hears it, it waits for the producer thread to
end, then terminates.

### `nexus writer`

When the signal is received, the main (and only) thread blocks, moves
any nxs files in the `move_cache` queue, then terminates.

## Instruction for review/testing

General code review.

Tested on simulated data.

Closes
https://github.com/STFC-ICD-Research-and-Design/supermusr-data-pipeline/issues/270.