Propagate incomplete frame data (#295)
## Summary of changes

Writes the `complete` flag of each frame into the nexus file, as well as
recording the digitisers present as a run log, in the case of `complete
== false`.

Also modified the type of the time field in the runlogs: `i32` changed
to `u64` as `i32` does not have the range for recording `ns` precision.

## Instruction for review/testing

General code review.
Has been tested with simulated data.

Please excuse the use of `anyhow`, will be changed later.