Set open telemetry levels via an environment variable (#318)
## Summary of changes

Changes the way we set OpenTelemetry levels and tidies up some spans.

- Removed `otel_level` cli option from all components.
- Modified OpenTelemetry filter layer to use environment variable
"OTEL_LOG" to determine which level to filter crates/modules on.
- Removed references to the `otel` target and `target = "otel"` syntax
in spans/events (This accounts for most files changed). This target is
not really useful, though OpenTelemetry-only targets may be introduced
in the future.
- In `trace-to-events`, functions processing individual channels moved
to their own file `channels.rs`. This allows the observability level for
these to be set via including `trace_to_events::channels=????` in
`OTEL_LOG`. At present use `info` to include channels, and
`warn/error/off` to exclude.
- Level of `spanned_root_as` functions changed to `trace` from `info`.
- Added `digitiser_id` tag to `Digitiser Event List` span in digitiser
aggregator.

## Instruction for review/testing

- General code review
- Has been tested on simulated data