Route se logs and run logs correctly (#328)
## Summary of changes

Sample environment logs can be delivered using both `f144_logdata` and
`se00_data` schemas. The current nexus writer identifies SELogs by its
schema, regardless of which topic it's delivered on. This causes SELogs
to be written as RunLogs.

This PR causes the writer to identify each type of Kafka message by the
topic it's delivered on, validate it's schema, and process it
appropriately.

- Topics are stored in a new struct `Topics` which is passed by
reference to functions.
- The `process_payload` function is refactored into functions which
handle each topic separately.
- The functions referred to in `process_payload` are moved into a
separate file.
- A `SampleEnvironmentLog` enum is introduced to handle the different
schemas that can be used.

## Instruction for review/testing

General code review
Tested with HiFi
(concise and correct instructions to aid the reviewer in verifying that
the changes work as intended)