Enable zstd compression (#349)
## Summary of changes

Provide the zstd library and enable zstd compression by default in all
components that interact with Kafka.

This is partly experimental to see what the effects are on compute
requirements and network throughput are for data "inside" the pipeline.

## Instruction for review/testing

- Run a demo pipeline, using the container images, with and without
these changes and see that the same output is still obtained.
- Verify that the data produced is actually compressed (can be done via
Redpanda Console).

---------

Co-authored-by: Modularius <daniel.kirk@rocketmail.com>