use filename for file name (#386)
## Summary of changes
Closes #382 

This makes the nexus-writer use the `filename` attribute of a `pl72` run
start message, rather than the run_name.

## Instruction for review/testing

send a pl72 run start with `filename` filled in and see that it
generates a file named accordingly. tested on my machine but probably
worth a manual test on someone elses.