Additional metrics in event formation (#394)
## Summary of changes

Adds more metrics to event formation that were previously in
`trace-telemetry-exporter`, removing the need for this as a separate
component (which has also been removed in this PR).

Only the metrics that proved useful have been moved. Message rate can be
derived when querying Prometheus or InfluxDB so has been excluded.

Closes
[#320](https://github.com/STFC-ICD-Research-and-Design/supermusr-data-pipeline/issues/320).

## Instruction for review/testing

- Run `trace-to-events` with simulated data
- Inspect the generated metrics via `curl`