Server path (#406)
## Summary of changes

Fixes the issue preventing us hosting on a subfolder via reverse proxy.

~~CLI parameter `server-path` changed to `server-url`, and now expects
the absolute URL the app is hosted at.~~
New CLI parameter `server-origin` expects the origin hosted at. Along
with `server-path` (which now requires a preceding slash), this
specified the full URL the app is hosted at.

~~I've had to dispense with the `<Router>` approach, which affects
in-site navigation (we only expect to navigate to /help but help could
be provided in a dialog box instead, so not a huge issue). I will try
fix this as a future issue.~~

## Instruction for review/testing

- General code review.
- Test MuSR deployment with parameter `--server-path=trace-viewer`
replaced with `--server-path=/trace-viewer` and with the added argument
`--server-host=https://ndw2982.isis.cclrc.ac.uk`