make remove_prefixes strip off block prefix but leave other PVs as full addresses (#419)
## Summary of changes

Closes #418 

All IBEX blocks are prefixed with `IN:INST:CS:SB:x` and these should end
up in a nexus file as `x`, but we have some other stuff in the
`_sampleEnv` topic which are archived values - these don't really _need_
to be in the nexus file but there's no harm them being there. Previously
if you stripped at the last colon for:

`IN:IMAT:SOME:VALUE:SP`
and 
`IN:IMAT:SOME:OTHER:VALUE:SP`

You'd end up writing to the `SP` dataset which isn't correct. 

This PR either strips off the prefix if the PV is prefixed with
`:CS:SB:` or just returns the full name if not.

## Instruction for review/testing

Test with some `f144` blobs with the above name.